账户余额导入Excel

+++ balance-vue/src/views/sys/balanceDetail/index.vue
@@ -189,9 +189,20 @@ export default {
         Message.error('请选择要删除的记录')
         return
       }
-      deleteAction('/sys/balanceDetail/delete', { ids: ids.toString() }).then(({ msg }) => {
-        Message.success(msg)
-        this.list()
+      this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteAction('/sys/balanceDetail/delete', { ids: ids.toString() }).then(({ msg }) => {
+          Message.success(msg)
+          this.list()
+        })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        })
       })
     },
     selectionChange(selectedRecords) {
